+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: jjourdan <jjourdan@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 02:00:45 by jjourdan          #+#    #+#             */
-/*   Updated: 2020/11/23 02:00:45 by jjourdan         ###   ########lyon.fr   */
+/*   Updated: 2020/11/24 22:08:31 by jjourdan         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strdup(const char *s)
 {
 	char	*dup;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	if (!(dup = malloc(sizeof(*s) * (ft_strlen(s) + 1))))
