+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: jjourdan <jjourdan@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 02:01:02 by jjourdan          #+#    #+#             */
-/*   Updated: 2020/11/23 02:01:02 by jjourdan         ###   ########lyon.fr   */
+/*   Updated: 2020/11/24 22:18:01 by jjourdan         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	int	i;
+	size_t	i;
 	int	size_src;
 
 	i = 0;
 	size_src = 0;
 	while (src[size_src] != 0)
 		size_src++;
-	while ((i < (int)size - 1) && (src[i] != 0))
+	while (((int)i < (int)size - 1) && (src[i] != 0))
 	{
 		dest[i] = src[i];
 		i++;
